GUI: Refactored the code that generates export files and the simcraft execution command.  They were previously parallel implementations of the same concept: now they both employ a common function for generating the simcraft parameters from user input.
Also, because not-checked check boxes in HTML simply don't show up in the form submission, not-checked options were being left out of the simcraft command, resulting in the simcraft defaults and not necessarily '0' as the user intended.  This refactor results in explicit values for ALL attributes, so there should no ambiguity (though exported files will bea bit bigger).


git-svn-id: https://simulationcraft.googlecode.com/svn/trunk@2436 3b4652a1-8050-0410-ac47-3d40261b0f8b