Initial import of the DBC extracted data along with a basic access API. It's somewhat messy and could use some definite prettying up and refactoring but it's a base to work off.
The actual imported data is in:
sc_spell_data.inc: Spell and Effect data.
sc_talent_data.inc: Talent data.
sc_scale_data.inc: Scaling info and ratings etc.

The hand-done values are in:
sc_extra_data.inc

The access struct is sc_data_access_t which is declared in sc_data_access.h and defined in sc_data_access.cpp

It's a child class of sc_data_t (from sc_data.h/sc_data.cpp)

In the constructor to sc_data_access_t or sc_data_t you can optionally pass in a pointer to an existing class and it will set that as the "parent" which means it'll copy the data from the parent initially and upon a reset.
The aim for this is so you can change values at different levels. Say on a global level, a sim level, a class level and a player level via options or profile settings etc.

There's still a lot of work I want to do with this stuff including implementing all the relevant spell effect types which will entail some big changes to how we do things I imagine however, in the meantime if you're wondering about the values of spell coefficients or min/max values, that data is available....

git-svn-id: https://simulationcraft.googlecode.com/svn/branches/cataclysm@4660 3b4652a1-8050-0410-ac47-3d40261b0f8b