To ensure that we actually can access the PTR data files I needed to move some things around.
Now if you're calling new talent_t() or active_spell_t() or passive_spell_t() it shouldn't be done in the player class constructor, but instead in either init_talents() or init_spells().

The talents you add should be placed in init_talents() and before the call to player_t::init_talents() (to ensure that the ranks are parsed correctly).

The problem before was the input parser creates classes as it finds them which means they can be created before the patch=<whatever> value might be set in which case we couldn't know which set of DBC files to be using.

The point being, the player class constructors shouldn't assume anything about the state of DBC data. They will have the correct DBC set when init_data() is called at the start of the init() method.

git-svn-id: https://simulationcraft.googlecode.com/svn/branches/cataclysm@4937 3b4652a1-8050-0410-ac47-3d40261b0f8b