1) Added some enchants and meta gems. 2) Added some potions etc. 3) Added the ability to provide a: type=cloth/leather/mail/plate in the item string. player_t will automatically determine whether or not the player gets to benefit from the 5% primary stat bonus or not based off this. For now if there's no type= field it assumes it's a "matching" type. It also filters out of the slots where it doesn't apply.
How to get this to work for your class: In your class constructor set: matching_gear_type = STAT_INTELLECT; (or equivalent).

Be sure to remove your other overrides like I removed mysticism=1 from the Priest module.

4) Also added a addon= gear option which is intended to function like a second enchantment on the piece of gear to support things like engineering enchants which can stack with normal enchants.

5) Also added a reforge= gear option which is intended to function like a modifier to the stats and only supports a limited range of stats.

As the various external sites don't fully support things like reforging and 2nd enchants there's a few TO-DO comments in the code but it returns sane info in the meantime.

git-svn-id: https://simulationcraft.googlecode.com/svn/branches/cataclysm@5123 3b4652a1-8050-0410-ac47-3d40261b0f8b