stat_type matching_gear_type replaced with:
virtual double matching_gear_multiplier( const stat_type attr ) SC_CONST { return 0.0 };

I've added method overrides for each class to the (hopefully!) correct thing for each.

Basically now it checks the attribute passed to it and depending upon buffs or primary_tree() or whatever it returns the correct multiplier.
The Warlock one uses the value from the nethermancy spell (which is a good thing). For the rest they're just hard-coded at 0.05 for now if it matches.

Over-time each module should be done more like the Warlock version so the value automatically updates if Blizzard change it around.

git-svn-id: https://simulationcraft.googlecode.com/svn/branches/cataclysm@5140 3b4652a1-8050-0410-ac47-3d40261b0f8b