Reforging now follows the format:
reforge=<stat_to_reduce>_<stat_to_gain>

i.e. reforge=spi_haste

This will remove 40% (rounded down) of the spirit on the item and give it as haste.

It will give an error if you don't specify 2 stats (unless you specify none of course) or if the type of stats aren't valid reforging stats or if the stat you're reforging from doesn't exist on the item or the the stat you're reforging to already exists. (i.e. Just like it functions in game).

git-svn-id: https://simulationcraft.googlecode.com/svn/branches/cataclysm@5185 3b4652a1-8050-0410-ac47-3d40261b0f8b