A very initial implementation of a spell data searching system.
Spell data searches can now be done by using the "spell_query" option in
simc. This does not do any simulation, but rather reads the expression given as
a value for the option, and runs it against the spell (or talent) data for
patch level that can be also specified on the command line e.g. "patch=4.0.3"
for cata. The result is a textual output of all spells (or talents) matching the 
given query expression.

To accomplish this, a new expression data type has been defined
(TOK_SPELL_LIST), which contains a list of spell or talent identifiers. To this
end, spell_query works on an expression level by first taking a full list of
spell ids, and performing filtering on that list to limit it. Thus, all
spell_query related operations are binary, and the left side operator must
always be a TOK_SPELL_LIST operand. Binary operators may be chained normally,
as the binary operator inherits both data source and the result type from the
left side operator.

The expressions for spell query follow the normal simc expression input,
however the spell query must be told where to get the initial list of
identifiers. The following identifier data sources have been defined:
  * spell - The master spell list, containing all spell data in simc
  * talent - The master talent list, containing all talent data (NOTE: not spells)
  * talent_spell - All spell ids used by the master talent list for talent ranks
  * class_spell - All spells belonging to a class and "clickable", this includes pets
  * race_spell - All racial spells
  * mastery - All class masteries
  * spec_spell - All specialization spells, passive or active
  * glyph - All glyph spells
  * set_bonus - All set bonus spells

A data source can be filtered by giving it a data field name, by suffixing the
data source name with a period and the filtering data field name. Currently the
system does not support deeper "paths" than 2, i.e. "spell.name" is the deepest
it will go. The expression is then given an operand, e.g. == !=, <, > and so
on, after which a string, numeric type or another spell identifier list (if the
operands &, | or - are used). The following data field names are currently
supported, listing the name, data_source_types (spell covers all but "talent"
data source), operand_type and a brief description: 
 * name, spell/talent, STRING
 * id, spell/talent, NUMBER
 * flags, spell/talent, NUMBER (not used for anything currently)
 * speed, spell, NUMBER (projectile speed)
 * school, spell, STRING (spell school name)
 * power_type, spell, NUMBER (spell resource type)
 * class, spell/talent, STRING (class name)
 * pet_class, talent, STRING (pet talent tree name)
 * scaling, spell, NUMBER (spell scaling type, -1 for "generic scaling", 0 for 
   no scaling, otherwise class number)
 * level, spell, NUMBER (spell learned level)
 * min_range, spell, NUMBER (minimum range in yards)
 * max_range, spell, NUMBER (maximum range in yards)
 * cooldown, spell, NUMBER (spell cooldown, in milliseconds)
 * gcd, spell, NUMBER (spell gcd duration, in milliseconds)
 * category, spell, NUMBER (spell cooldown category)
 * duration, spell, NUMBER (spell duration in milliseconds)
 * cost, spell, NUMBER (spell resource cost)
 * rune, spell, STRING, (b = blood, f = frost, u = unholy, will match minimum 
   rune requirement)
 * power_gain, spell, NUMBER (amount of runic power gained)
 * max_stack, spell, NUMBER (maximum stack of spell)
 * proc_chance, spell, NUMBER (spell proc chance in percent (0..100))
 * initial_stack, spell, NUMBER (initial amount of stacks)
 * cast_min, spell, NUMBER (minimum cast time in milliseconds)
 * cast_max, spell, NUMBER (maximum cast time in milliseconds)
 * cast_div, spell, NUMBER (scaling divisor for cast time, always 20)
 * desc, spell, STRING (spell description)
 * tooltip, spell, STRING (spell tooltip)
 * tab, talent, NUMBER (talent tab number, 0..2)
 * dependence, talent, NUMBER (talent id this talent depends on)
 * depend_rank, talent, NUMBER (talent rank of talent id this talent depends on)
 * col, talent, NUMBER (talent column 0..3)
 * row, talent, NUMBER (talent "tier" 0..6)

For numeric data fields, the following numeric operators between a spell list
and a numeric right-side operand are available: ==, !=, >, <, >=, <=. All string data 
fields can be filtered by a string right-side operand and the following operators:
 * == : case insensitive string equality
 * != : case insensitive string inequality
 * ~  : case insensitive substring in data field
 * != : case insensitive substring not in data field

**IMPORTANT NOTE** Due to how simc does command line parsing, all string based
filtering needs to be input with whitespaces converted to _ and without any
special characters. The spell filtering will automatically "canonicalize" any
strings it compares your textual input against, and all comparisons are done
case insensitive. This means that for example "Shadow Word: Pain" becomes
"Shadow_Word_Pain" when given as a right-side string operand to filter with
**IMPORTANT NOTE**

Furthermore, for identifier list operands on the right side of the expression, the 
following operators are available:
 * &  : intersection (identifiers in both left and right side)
 * |  : union (identifiers from both left and right side, removing duplicates)
 * -  : subtraction (right side identifiers removed from left side list of identifiers)

Examples:
  a) All activatable class spells of shaman
     $ simc spell_query=class_spell.class=shaman
  b) All cataclysm activatable spells
     $ simc spell_query="class_spell.level>80"
  c) All activatable class spells of shaman and priest
     $ simc spell_query="class_spell.class=shaman|class_spell.class=priest"
  c) All mastery spells
     $ simc spell_query=mastery
  d) All class (and pet) spells costing resources
     $ simc spell_query="class_spell.cost>0"
  e) All spells that contain "shadow damage" in their description or tooltip
     $ simc spell_query="spell.desc~shadow_damage|spell.tooltip~shadow_damage"
  f) All death knight spells that use at minimum a blood and an unholy rune
     $ simc spell_query="spell.class=death_knight&spell.rune=bu"
  g) All spells using the constant scaling
     $ simc spell_query=spell.scaling=-1
  h) All spells that are fire or frost based (including derivations)
     $ simc spell_query="spell.school=fire|spell.school=frost"

Note that depending on your shell interpreter, you may need to escape some special
characters used in the spell queries.



git-svn-id: https://simulationcraft.googlecode.com/svn/branches/cataclysm@5206 3b4652a1-8050-0410-ac47-3d40261b0f8b