Fix Power Torrent and add in all the "easy" blue and epic non-PvP DPS trinkets.
More difficult trinkets will have to wait for Nate to implement the mechanics we need for them.

Almost all of these trinkets added have placeholder ICD values.
I've included a "// TO-DO: Confirm ICD" comment on each one with an ICD. As they're tested this can be removed.
I imagine a lot of them will follow the same pattern of ICD but we'll see...



git-svn-id: https://simulationcraft.googlecode.com/svn/branches/cataclysm@5449 3b4652a1-8050-0410-ac47-3d40261b0f8b