Add a batch file and a UNIX script to extract the DBC data from the DBC files.
Requires Python to be in the path and the user to specify a directory with the full DBC files.

Usage: generate [ptr] <build> <inpath>

Example:
To generate the PTR versions of build 13316 when all the DBC files are in d:\data\code\dbc\dbc\13316\DBFilesClient:

generate ptr 13316 d:\data\code\dbc\dbc

(note that DBFilesClient is appended by default).

The output will be placed directly into the engine directory so long as these scripts are kept in the dbc_extract directory along with dbc_extract.py.

git-svn-id: https://simulationcraft.googlecode.com/svn/branches/cataclysm@5542 3b4652a1-8050-0410-ac47-3d40261b0f8b