Following issue 521 : New "proxy" global option added 	syntax  : proxy=type,host,port 	example : proxy=http,proxy.mywonderfulcorp.com,3128 	example : proxy=http,10.120.45.2,8080
The "proxy" option MUST be provided BEFORE any other "web-import" options (armory, chardev...) in scripts and command-line
	good : ./simc proxy=http,proxy.example.com,3128 armory=eu,dalaran,kartov
	bad  : ./simc armory=eu,dalaran,kartov proxy=http,proxy.example.com,3128

Only transparent, unauthentified http proxies are supported (squid and the likes) so the "http" type is the only option for now

For now, the option is effective for POSIX codebase only. It is completely ignored for MinGW/VC++ versions, and i'll try to work on these.



git-svn-id: https://simulationcraft.googlecode.com/svn/trunk@7968 3b4652a1-8050-0410-ac47-3d40261b0f8b