PTR Build 14288 DBC Data.
Remove the ptr overrides that were waiting on this patch.

Note that the +dmg for Shred and Ravage! do NOT match the Patch Notes. They seem to be the same value as the base damage of Rake. This is a bug on Blizzard's behalf. Testing will be needed to see what it's actually doing in-game.

git-svn-id: https://simulationcraft.googlecode.com/svn/trunk@8681 3b4652a1-8050-0410-ac47-3d40261b0f8b