Move chart_t into the report part of SimC, decouple it from the main global namespace.
sim_t::report_information_t and player_t::report_information_t are
intended to contain the neccessary report information, and I'll probably
extend this even further and create them for other classes as well.

One goal here is to pass if possible only the absolutely necessary
information to a report function, without it doing whatever it will with
a player or sim pointer.

There have also been a lot of const char* to std::string changes, and I
added a color namespace similar to the one we have in simc-swtor.

This revision throws a std::logic_error when reporting the html and
needs to be fixed.

git-svn-id: https://simulationcraft.googlecode.com/svn/branches/mop@11281 3b4652a1-8050-0410-ac47-3d40261b0f8b