Extend buff_creators, remove all but one buff_t constructor.
The original idea was to make eg. stat_buff_creator_t inherit from
buff_creator_t, and thus allow a easier set up for those derived buff
types. Unfortunately that does not work since the functions ( eg.
max_stack(...) return a reference to the base_class and not the derived
one ).
I decided to go with the non-inheriting way for now. So a
stat_buff_creator requires a buff_creator_t to work. If someone has a
better solution ( still allowing for buff-generation in one line ), feel
free to propose it. I know template magic could be used, but that seems
overly complicated.

git-svn-id: https://simulationcraft.googlecode.com/svn/branches/mop@11363 3b4652a1-8050-0410-ac47-3d40261b0f8b