Ignite Mechanics:
Update the whole mechanic to how it behaves in MoP.

Munching and Rolling are gone, but there is still a aura delay present.
Ignite damage is calculated after this delay, and a unmodified tick may
go off between the Ignite impact and it refreshing the dot.

Move the division by the number of ticks from the ignite trigger to the
ignite action.

Implement Priest Echo of Light as a Ignite-Like Spell.

git-svn-id: https://simulationcraft.googlecode.com/svn/branches/mop@12514 3b4652a1-8050-0410-ac47-3d40261b0f8b