Update caster profiles to use Relic of Yulon.
Mage: The mage profiles are a mess. They're completely hardcoded, not even using item id's and the default actions don't work.
Please, update the *code* to have it generate the actions.
Added item id for Essence of Terror in Mage profiles so that it at least used the same ICD we're using for other classes...

git-svn-id: https://simulationcraft.googlecode.com/svn/branches/mop@13611 3b4652a1-8050-0410-ac47-3d40261b0f8b