Implement Proxy Stormlash totem. New option "override.stormlash" controls it's use. 0 = no stormlash, > 0 = that many stormlash totems used.
The proxy system will use the stormlash totems back to back, with one second delay between the "casts". If the bloodlust in the simulator is configured to be used in the first 30 seconds of the fight, the first set will be synced to bloodlust. Finally, once the all the proxy totems have been used, there is a 300 second delay (starting from the first proxy totem) until another series of totems will be used.

Shaman:
Enable stormlash totem in default action list.
Regenerate all profiles.

Implement an option (aggregate_stormlash=1/0, default=0) that combines all stormlash totem damage to the Shaman who cast the stormlash, in addition to including the DPS of the stormlash effect for each individual actor. Requires that override.stormlash=0. If for some reason two shamans oveerlap their stormlash totems in the simulator, the Shaman who first cast the totem will get the contribution. 

Rogue:
Optimize rogue_melee_attack_t::ready()



git-svn-id: https://simulationcraft.googlecode.com/svn/trunk@14206 3b4652a1-8050-0410-ac47-3d40261b0f8b