Support PREFIX and DESTDIR properly in command line Makefile, and in command line SimulationCraft.
In essence, the command line Makefile sets PREFIX automatically to "..", and during compilation PREFIX is exported to the code as SHARED_DATA define. The options system automatically adds the default paths (current directory, and parent directory) to the profile search path, but in addition if SHARED_DATA is not equal to "..", adds SHARED_DATA directories to the search path as well. By default, nothing has changed, and all files go to parent directory with make install, as before.

In addition, if PREFIX/DESTDIR are set (differs from ".."), the Makefile now copies our profile directory structure to $DESTDIR$PREFIX/share/SimulationCraft by default, and the binary ("simc") to $DESTDIR$PREFIX/bin.