Implement experimental ticking buffs.
buff_creator_t (and friends) have several new parameters to control the ticking.
- period() that sets the tick time in seconds
- tick_behavior() that sets how the buff acts on refreshes (BUFF_TICK_CLIP, BUFF_TICK_REFRESH), also BUFF_TICK_NONE can be used to disable a ticking buff. Otherwise the behavior is the same as with current dots (clip restarts the ticking process, refresh allows ongoing tick to finish, discards rest).
- tick_callback() a std::function<void(buff_t*, int, int)> function callback that is called on-tick, immediately before the buff's aura state is altered. Parameters are 1) the buff object 2) current tick number (1..TOTAL), 3) total number of ticks, counting from the last start time of the buff.

Additionally, buff construction now automatically detects the periodicity of the buff from spell data, however to enable the ticking behavior, buff_creator_t::period() must be used to explicitly set the behavior of the buff.

Currently no support for: immediate ticks, or hasted ticks.

Under the hood, the ticking system does not "trigger" the buff every period seconds. Instead, it calls lower-level "bump", and "decrement" functions to alter the aura state. There may be issues with ticking buffs and reporting due to this. As a consequence of this change, the "refreshes" parameter from buffs is also removed. A ticking buff is considered to be "non refreshing", and everything else in simc refreshes normally, when triggered while being up.