Added touch of elune racial.
All it does it add 1% crit/haste to the current highest stat rating on the player. Eventually we can add a player option to select night/daytime, I guess.
Won't work if the player manages to have exactly equal stat ratings... not sure what type of tie breaker to use.