Introduce an experimental "strict_sequence" action.
Same syntax as normal "sequence" action, but with a twist. Strict_sequence requires that all the actions are "ready" before strict_sequence executes. In turn, once the actor has committed to a strict_sequence, it will perform the sequence through, before returning to the normal APL behavior.

Some caveats apply:
- If there are actions that have inter-dependencies inside the sequence (i.e., a subsequent action becomes !ready due to an earlier action executing), the strict_sequence will cancel
- Any "interruption" (stun, movement etc) will cancel/reset the strict sequence
- There is no way to restart a strict sequence, once you have committed to it, it will be run through (or canceled, see above)