Resolve: ignore temporary max health buffs. NOTE: this is predicated on all max health buffs passing a true temporary_stat flag to player_t::stat_gain(). If you are hardcoding a gain (as in Last Stand... ALEX!) and want it to count as a temporary buff for the purposes of Resolve, you need to make sure it passes that argument, otherwise it will default to false and will negatively impact your Resolve experience.
Also note that we kludged this by storing temporary max health gains in temporary.resources[RESOURCE_HEALTH]. This is a bit of a kludge, but it seemed better than adding more fields to gear_stats_t. If anyone notices a reason this will cause a problem, let me know and we'll un-kludge it somehow.