Implement "variable" action, for generic variable store. Options: name: Mandatory. The variable name. op: The operation, 'set', 'print', 'reset', 'add', 'sub'. Default 'set'. value: The value parameter for 'set', 'add', 'sub'. Mandatory for those operations, unused otherwise. Uses normal "if" expression syntax. default: The default value of the variable. Default 0. delay: The cooldown of the action in seconds. Default 1 seconds for 'print', 0 for other operations.
The action will not execute in the following conditions:
If operation is 'set' and the value evaluation is the same as the current value of the variable.
If operation is 'add', or 'sub', and the value evaluation returns zero.
If operation is 'reset', and the variable value is already default value.

Variable values can be used in expressions using the 'variable.foo' expression, where 'foo' is the variable name.