Add a second dot copying behavior. dot_copy_e enum chooses the copy type.
DOT_COPY_START is the default (and original) behavior, where the dot state is copied from the source to the target, and a completely new dot is started.
DOT_COPY_CLONE is a new behavior, where the full dot state is copied from the source to the target, including remaining time, and if the copy is performed mid-tick, the remaining time-to-tick for the next tick.

Also fixed some potential bugs in the copying code.