Stat calculations: Corrected bug in how we handle stats.
Attributes (e.g. Strength) are not actually integers in-game. They are doubles, which then get floor()-ed at various points (for example, to display on the character sheet). Based on testing with the 5% stat buff on paladn/warrior premades, we've determined that our existing formula gave a consistent error of -1 Strength compared to the character sheet. This is consistent with the base attribute not being floored after applying the matching gear multiplier.

In short, for a matching multiplier m (=1.05) and kings multiplier k (=1.05), our old formula was (condensing several pieces of code together):

total_stat = floor( ( floor(gear_stat * m ) + floor( base_stat * m ) ) * k )

The corrected formula is

total_stat = floor( ( floor(gear_stat * m ) + base_stat * m ) * k )

So far, this new formulation agrees with in-game results in 14+ cases in which the old one did not.

For completeness, it is not even clear that the outer floor() is performed to get total_stat; it is entirely possible that it is only done when displaying the value on the character sheet. However, we have no easy way of testing this from the character sheet alone, obviously. It may be that the util::floor() in player_t::get_attribute() is in error as well, though the practical impact on simulation output is insignificant.