[Death Knight] Add "dot.necrotic_plague.stack" expression to return the number of stacks. Add "disease.<dot_expression>" shorthand to check the disease dots on target.
dot_expression can be any dot expression (ticking, remains etc), in which case each diasease (FF, BP, NP) is checked, and the first non-zero result will be returned.

In addition, <dot_expression> can also be prefixed with a "min_ or "max_" (e.g., min_remains), in which case all available diseases are checked, and the minimum or maximum value will be returned.