Resolve: Moved resolve timeline into its own structure, fixed reporting/accumulator bugs.
Previously, every time resolve was recalculated it was being added to the accumulator. This meant that if 2+ update() calls happened within a second (which happens often), it would double (or triple, or etc.) the value stored in the accumulator.

To fix this, we need to be able to check to see whether we've already recorded a resolve value in the element of the array before we write to it, and if so, store some modified value. I've chosen to store only the maximum (see the resolve_timeline_t::add_max() function), but if we were willing to store more data we could use a weighted average.

Note that this precludes use of a single sc_timeline_t for Resolve, since we need to clear the timeline every iteration. The resolve_timeline_t structure has two timelines to facilitate this, similar to the health_changes_timeline_t structure.