sim_t::combat_end(): execute datacollection_end() on iteration 0.
While working on the previous commit, I noticed that we are collecting data on iteration 0, but not running datacollection_end(). Thus, the divisor timeline we're using is one smaller than it should be (example: if iterations=10, the max value in the divisor timeline is 9). This exaggerates all of the timelines in the sim, though this is most noticeable in the resource timelines (Paladins having 6 holy power, warriors having 150 rage, etc.).

Since I'm not sure what other repercussions this particular change has, I'd like serge and navv to review it.