Add an expression to evaluate on-use buff stat.
The expression "use_buff.<stat_name>" used in conjunction with the "use_item" action evaluates to
true (1) if the slot/item in question (specified by the 'name', or 'slot' parameter) has a stat
buff of type <stat_name>. Otherwise, the expression always evaluates to false (0).

Note that this expression only functions in conjunction with the "use_item" action.