Fix an once in a million issue with dots.
Mass-extended dots could go into a state that flips the dot end event, and dot tick events. This
caused a sanity-check assert to trip. Code changed so that dot end event only performs the last
(damaging or healing) tick if there are ticks to do. Add a different sanity assert to ensure that
if the rare tick-before-end-event occurs, both of the involved events occur on the same timestamp.