[Warrior] Add new option - non_dps_mechanics - Default value is false. When set to false, this disables spells like bloodthirst heal, impending victory heal, second wind, blood craze.
This option is automatically set to true for tanks.

Also, set use_off_gcd to false for charge, after profiling the warrior specs I noticed that charge_t::ready was the #10 most commonly used function in the entire simulation. This will reduce that quite a bit.

The combination of these two will speed up warrior simulations.