Allow spell_query level to go past MAX_LEVEL.
If this happens, spell_query presumes that the spell or spells you print
are trinket spells, and scales the effect (Scaled Value) by using the
item budget of the ilevel you specify with the level. If you use this
for anything other than item-related spells, things will look weird.
Buyer beware.