Implement Insatiable Hunger.
Note that there's currently an unknown behavior involved with the
trinket. The driver has an ICD of 10 seconds, while the buff has a
duration of 20 seconds. This means the effect can re-proc while the
stacking phase is ongoing. The sim currently treats a re-proc as just
another added stack, but there are other possibilities, namely:

1) Reset the buff to 1 stack, refresh duration
2) Do nothing ("more powerful spell is already up")
3) Add a stack (or don't add a stack) and refresh duration