Fix an issue with pet-related expressions.
Before the pet init changes, pets were always created, and as such, any
"pet.foo.bar" expression would create the actual expression. After pet
init changes, pets are no longer created and it is possible that the pet
in the expression ("foo") is not found. To avoid user headaches of
having to edit APLs when pets related to for example talents are not
there, the system now returns a constant expression if the pet is not
found (instead of not failing to recognize the expression).

The value of the constant was set to 0, which unfortunately triggers
expressions like "target!=pet.prismatic_crystal", as 0 is a valid
(enemy) actor id, usually assigned for the first target. The value was
changed to -1 in this patch, arguably not any better, but knowing that
99.9999% of the simulator expressions return positive values, should be
safe enough.