Add in support for ground aoes with fancy_target_distance_stuff. Basically, any spell that falls under these rules, will be considered a ground-targeted aoe.
1: Uses tick_action or add_child
2: Has a range AND radius over 0 yards. Radius is used as the radius of the aoe.

How it works:
During initialization, the sim tags children and tick_actions with the name of the dot driver.

During runtime, the parent dot calls start(duration) and original_x and original_y are marked as the location of the target when the ground aoe was first cast. When the child ticks/executes, it checks to see if there is a parent dot. The children use the original_x and original_y values for determining if targets are still in the aoe.

I had to change how a couple of spells were setup to function properly with this, notably explosive trap and meteor.