Change default dot behavior for all spells to be DOT_REFRESH, as that is the behavior for nearly every spell nowadays. It's much easier to override spells to dot_clip than the other way around.
Add a new function to target distance, which assists target_if by removing targets that it cannot hit in the first place.

Fix a bug with prismatic crystal in select_target_if_target which wasn't adding the crystal to the target list. Also, fix this with distance targeting by making sure that distance targeting goes through the normal target_list first, and then deletes any targets that are out of range later.

Add more log messages, and allow the sim's main target to be moved.