Fix an issue with heirloom scaling.
The heirloom upgrade system is implemented through bonus ids. The base
item data always includes a "base scaling system", that is used when the
heirloom is not upgraded. This commit adjusts the behavior in simc so
that if a scaling bonus id is given for a heirloom, the "base scaling
system" is not used.