Fix another issue with spell data overrides.
It turns out that Blizzard re-uses spells in multiple driver spells.
Previously, simc presumed that there is a 1:1 mapping between a trigger
spell and a driver spell. This commit adjusts that behavior so each
trigger spell may have multiple drivers, and as a consequence, fixes the
issue where overrides on triggered spells with multiple drivers were not
performed correctly.

The commit also includes "Triggered By" information to textual spell
query.