Fix an issue with target_if.
If target_if was used with the "give me first available target", the
target_if expression was never evaluated for single target simulations.
This made it so that target_if=!ticking for example would allow a dot to
be cast on the (single) target, even if it was ticking.

Now, target_if is always evaluated for "first" mode, even in single
target scenarios. The behavior for "min" and "max" modes are like before
(returns the only available target).