Adjust item bonus handling in relation to ilevel changes
Now, if an item only includes a "base ilevel" bonus, it will be used to
set the ilevel of the item. If it includes any other ilevel adjusting
bonuses, they will be used instead.

There is strong indication that this is not exactly how things work in
game, however it should not break anything, and it will help in creating
sample profiles for us, since we can only tack on the base ilevel
setting without having to hunt for a suitable "adjust ilevel" bonus to
actually increase the ilevel.