Fix DH memory leak, and globally some queueing related issues
Some abilities require a specific buff to be up to be able to execute.
Since the introduction of the queueing system (where abilities are
queued before their cooldown has elapsed), there are tiny windows in
time where the queueing system allowed bypassing of the buff check.

The obvious cases have now been fixed to compare the buff remaining time
to the queue delay of the ability cooldown. This will ensure that the
ability will not be queued mistakenly, outside of the buff's up state.