Refactor event_t creation interface.
Add new event_t constructors accepting delta_time, which directly
schedule the event. For 95%+ of the use cases, those are the
constructors that should be used, as they ensure that we do net get any
unscheduled, but allocated events.

While we're at it, add bool scheduled; variable to safety-check against
double-schedules.

Hide placement-new operator behind a new make_event function. This
function also checks if the event has been scheduled, to avoid those
tricky-to-debug problems.

The placement-new operator is protected for now, because of death_knight
dynamic_event_t, without with it could be private.