Add a simple verification of use_item actions
The will now verify that all item-based use actions have a corresponding
use_item action created for them (i.e., there's an use_item line for the
item/slot in the action priority list). For the custom APLs, where some
items are deliberately not meant to be used, there is a new option
"use_item_verification=0" that will disable the verification (and the
warnings).