Add some naive targeting logic to special effect procs. Fixes #3180.
Now, DBC-based proc callbacks check the proc action for the special
effect in the case of incoming (e.g., damage taken) events. If the
action is a heal, the player itself is targeted, otherwise the source of
the event (e.g., the enemy) is targeted. This behavior is also
overridable for custom callbacks by overriding the
dbc_proc_callback_t::target method.