[Hunter] refactor Marked Shot
Marked Shot is pretty special due to the fact it is implemented as a server
side script. It's not even an "aoe" in the usual sense as it looks like the
server builds an internal list of valid targets and then processes it in order
with a slight delay between each with only the "first" target (which is
somewhat random, it's not the closest nor the current player target) processed
"instantly" on Marked Shot cast.
(In one testing with 11 targets there is alsmost a full second between Marked
Shot cast and last Vulnerable application).

This is a first step in trying to simulate this behavior by simulating the
"2-stage" nature of Marked Shot - we factor out the Marked Shot damage spell
from the "driver" spell.