[Priest] Rewrite insanity drain to a trigger-based system
Before, Insanity was drained every 50 milliseconds based on the actor
state. This commit implements state tracking for insanity, allowing
millisecond accurate granularity for the draining (and thus, the inanity
level of the actor).

The only visible change in reports will be different kind of average
drain amount (higher), and Dispersion and Void Torrent "insanity
savings" are not reported any more.