Experimentally implement Draught of Souls
This trinket is a bit of a special case, since it's actually triggering
a channeled thing, meaning we need to model that. The on-use system we
have is ill-equipped for that, so some jumping through hoops is needed.
It should work fine (and seemingly does), but if anything weird pops up
in conjunction with the trinket, likely reason is the weird "background
channeling" thing we need to do here.

Oh, and there's no support for the "immobile" thing for now, since simc
does not really understand that as a concept anyhow.