Adjust how the simulator computes DBC-backed enchant stats
It seems that Blizzard has stopped updating (through hotfixes) the
master enchantment data table for entries that are actual item enchants
(i.e., permanent enchants through the enchanting profession). As a
result, the simulator failed to detect the change for Mark of the
Trained Soldier enchant (300 -> 600 mastery rating). This commit changes
the behavior of the computation so that if a mapped spell id is
available in the composite enchant data table we have, it will generate
the enchant stats through the spell data, instead of the enchant data
table.

End result for the vast majority of enchants should be no change, but
Mark of the Trained Soldier will start working properly now.