Move critical damage bonus calculation to impact
Turns out critical damage bonuses are actually calculated on impact in
game, as in a buff that increases the actor's crit damage bonus while a
spell is in-flight, will actually affect it.

This should not break anything, but if some ability in a class module
suddenly loses crit damage bonus (crit = hit damage), it's likely due to
this.