[Hunter] fix cooldowns affecting legendaries
We can't modify action cooldowns in ::init() as its called for every
line in the apl with the given action but cooldowns are unique. Which
means that, if we get more than one given action in the apl, the
cooldown is going to be modified several times.

As we don't have legendaries initialized in action constructors move it
all to hunter_t::init_special_effects().

Does not affect our default apls as all of the affected actions are
present only once in them.