[Hunter] add T19 2pc BM hotfix
Implement 2017-02-06 T19 BM 2pc hotfix
  With the Dire Frenzy talent, the Eagletalon Battlegear Beast Mastery
  2-piece bonus should now grant your pet 10% increased damage for 15
  seconds.
Contrary to what the hotfix states the buff gets applied even *without*
Dire Frenzy talented.

For reference: https://goo.gl/pCzJ13
 the first main pet melee (47540)  at 0:35.502 is with BW up
 BW fades at 0:36.429-452 (note the 2 buffs fading from the main pet)
 the last main pet melee (32291) at 0:37.881 is without BW up
The hunter in question does not have Dire Frenzy. His BW should be 31% (no
Bestial Fury, no Unleash the Beast relics), but 47540 / 32291 is ~1.44
which is 44%. If we add in the 10% from the "second" BW it all lines up
(as 1.31 * 1.1 = 1.441 and 32291 * 1.31 * 1.1 = ~47540).