[Hunter] Spitting Cobra update
This thing is *very* special.

First, it fully inherits player damage multiplier (so it's buffed by
AotSkylord/Mongoose Power & artifact traits).

Second, it has brain/ability lag of around 340ms:

    N           Min           Max        Median           Avg        Stddev
 2146           0.0         805.0         421.0     341.03262     168.89531

With the actual timing being pretty inconsistent/strange, as it looks
like the lag *is* affected by haste effects, but in a non-linear way (in
some cases the lag under BL alone is lower than BL + Zerking combined).
There are also streaks of sub 50ms lag Spits in a row that are not
"connected" to anything (no haste buffs or anything).
[ to me it looks like the Cobra is driven by an "ai" with a set decision
  "tick" rate that has a somewhat random (or simply low) priority
  server-side ]

Model it through Cobra Spit ability lag (requires making the Cobra a PET
instead of a GUARDIAN it currently is, as the latter has ability_lag
"turned off") as simulating the full in-game behaviour is kinda
impossible due to it's inconsistency.