Rewrite Cinidaria the Symbiote
Cinidaria should now proc on everything else, instead of itself and
Legion enchants. Probably should disable other enchants as well?

If Cinidaria in the simulator is proccing an ability it should not, the
initialization of the special effect now contains a map of blacklisted
spell ids, where the offending ability can be added.