[Hunter] another pass on 7.2 artifact traits
Talon Bond, Echoes of Ohn'ara & Thunderslash should be fully working
now.

Cyclonic Burst is implemented as a simple aoe dot for now, in-game it's
basically a ground pulsing aoe from the Windburst trail (which is also
greatly affected by hitboxes). Considering we can't have non-circular
aoes this is closest we can get. May need to "nerf" it in the sim as
it's going to overvalue it.

Cobra Commander ideally needs more logs but should be pretty close to
in-game.