[Shaman] Add additional Storm and Fire Elemental pets
T20 set bonus will allow Elemental Shamans to potentially summon
multiple guardian versions of Fire or Storm Elementals.

In addition, this commit implements current in-game behavior for the
primal versions of the elementals, where if you attempt to summon one
while the primal elemental is already up, it will result in a wasted
cooldown.