[Demon Hunter] Fix a crash with the Fiery Brand ability
One place in the code did not check for the existance of the fiery brand
dot on the player, but implicitly used the pointer. This resulted in
crash for Havoc, if the user defined any kind of raid damage to occur as
a raid event (for exmaple the Ultraxion fight style).